old/pyproject.toml
[project]
name = "short-domain-find"
version = "0.1.0"
description = "A toolset to help you to find your favourite short domain."
readme = "README.md"
authors = [{ 'name' = 'wyf9', 'email' = 'wyf9@wyf9.top' }]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "colorama>=0.4.6",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "whois-api>=1.2.0",
]
